+++ src/value.h
@@ -9,8 +9,8 @@ typedef struct {
   // dynamic array containing zero or more values
   Value* values;
 
-  int len; // number of values in array
-  int cap; // available spaces for values in array
+  size_t len; // number of values in array
+  size_t cap; // available spaces for values in array
 } ValueArray;
 
 void init_value_array(ValueArray* arr);
